+++ crates/hwpforge-smithy-hwp5/src/layout_hint_patch.rs
@@ -189,6 +189,7 @@ fn collect_flow_paragraph_layout_hints(
             | Hwp5Control::Arc(_)
             | Hwp5Control::Curve(_)
             | Hwp5Control::ConnectLine(_)
+            | Hwp5Control::Equation(_)
             | Hwp5Control::OleObject(_)
             | Hwp5Control::Unknown { .. } => {}
         }
@@ -221,6 +222,7 @@ fn collect_scope_paragraph_layout_hints(paragraph: &Hwp5Paragraph, scope: &mut S
             | Hwp5Control::Arc(_)
             | Hwp5Control::Curve(_)
             | Hwp5Control::ConnectLine(_)
+            | Hwp5Control::Equation(_)
             | Hwp5Control::OleObject(_)
             | Hwp5Control::Unknown { .. } => {}
         }